refactor: replace Logo to components
